Extract Msf::DBManager::Import::Acunetix
MSP-11124
